feat: clitools and cli for B01_SL_load_single_file
